requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==1.2.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic==2.5.0
python-multipart==0.0.6 